requirements.txt
# Production-Grade Product Standardization Platform
# Python Dependencies

# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Database Connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
clickhouse-connect>=0.6.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP Requests (for Superset API)
requests>=2.31.0

# Logging & Monitoring (Optional but recommended)
structlog>=23.1.0
python-json-logger>=2.0.0

# Task Scheduling (Optional - for advanced automation)
schedule>=1.2.0
celery>=5.3.0
redis>=5.0.0

# Database Migrations (Optional - for advanced schema management)
alembic>=1.12.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality (Optional)
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
